refactor:extract GithubActionsPipelineCommitService from GithubActionsPipelineService
