+++ backend/src/main/kotlin/metrik/project/domain/service/githubactions/GithubActionsPipelineRunService.kt
@@ -7,15 +7,36 @@ import org.springframework.stereotype.Service
 @Service("githubActionsPipelineRunService")
 class GithubActionsPipelineRunService(
     private val buildRepository:BuildRepository,
-    private val githubRunService:GithubRunService
+    private val githubRunService:GithubRunService,
+    private val githubBranchService: GithubBranchService
 ) {
-    fun githubActionsRuns(pipeline:PipelineConfiguration): MutableList<GithubActionsRun> {
+    fun getNewRuns(pipeline:PipelineConfiguration): MutableList<GithubActionsRun> {
+        val latestTimestamp = getLatestBuildTimeStamp(pipeline)
+        var newRuns = githubRunService.syncRunsByPage(pipeline, latestTimestamp)
+        val branches = githubBranchService.retrieveBranches(pipeline)
+        return newRuns.filter { branches.contains(it.branch) } as MutableList<GithubActionsRun>
+    }
+
+    fun getProgressRuns(
+        pipeline: PipelineConfiguration,
+        newRuns: MutableList<GithubActionsRun>
+    ) {
+        val inProgressRuns = buildRepository.getInProgressBuilds(pipeline.id)
+            .parallelStream()
+            .map { githubRunService.syncSingleRun(pipeline, it.url) }
+            .toList()
+            .filterNotNull()
+
+        newRuns.addAll(inProgressRuns)
+    }
+
+    private fun getLatestBuildTimeStamp(pipeline: PipelineConfiguration): Long {
         val latestTimestamp = when (
-                val latestBuild = buildRepository.getLatestBuild(pipeline.id)
+            val latestBuild = buildRepository.getLatestBuild(pipeline.id)
         ) {
             null -> Long.MIN_VALUE
             else -> latestBuild.timestamp
         }
-        return githubRunService.syncRunsByPage(pipeline, latestTimestamp)
+        return latestTimestamp
     }
 }
